refactored naming for better conventions
