+++ src/main/java/interface_adapter/fetch_image/FetchImagePresenter.java
@@ -1,4 +1,4 @@
-package interface_adapter.image;
+package interface_adapter.fetch_image;
 
 import entity.CommonImage;
 import use_case.fetch_image.FetchImageOutputData;
