+++ src/main/java/interface_adapter/fetch_image/FetchImageRepository.java
@@ -1,4 +1,4 @@
-package interface_adapter.image;
+package interface_adapter.fetch_image;
 
 import entity.CommonImage;
 import java.util.List;
