+++ src/main/java/interface_adapter/fetch_image/UnsplashFetchImageInputBoundary.java
@@ -1,4 +1,4 @@
-package interface_adapter.image;
+package interface_adapter.fetch_image;
 
 import entity.CommonImage;
 import java.util.List;
