--- src/main/java/use_case/change_password/ChangePasswordInputBoundary.java
@@ -1,14 +1,0 @@
-package use_case.change_password;
-
-/**
- * The Change Password Use Case.
- */
-public interface ChangePasswordInputBoundary {
-
-    /**
-     * Execute the Change Password Use Case.
-     * @param changePasswordInputData the input data for this use case
-     */
-    void execute(ChangePasswordInputData changePasswordInputData);
-
-}
