--- src/main/java/use_case/change_password/ChangePasswordOutputBoundary.java
@@ -1,18 +1,0 @@
-package use_case.change_password;
-
-/**
- * The output boundary for the Change Password Use Case.
- */
-public interface ChangePasswordOutputBoundary {
-    /**
-     * Prepares the success view for the Change Password Use Case.
-     * @param outputData the output data
-     */
-    void prepareSuccessView(ChangePasswordOutputData outputData);
-
-    /**
-     * Prepares the failure view for the Change Password Use Case.
-     * @param errorMessage the explanation of the failure
-     */
-    void prepareFailView(String errorMessage);
-}
