--- src/main/java/use_case/change_password/ChangePasswordOutputData.java
@@ -1,24 +1,0 @@
-package use_case.change_password;
-
-/**
- * Output Data for the Change Password Use Case.
- */
-public class ChangePasswordOutputData {
-
-    private final String username;
-
-    private final boolean useCaseFailed;
-
-    public ChangePasswordOutputData(String username, boolean useCaseFailed) {
-        this.username = username;
-        this.useCaseFailed = useCaseFailed;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public boolean isUseCaseFailed() {
-        return useCaseFailed;
-    }
-}
