+++ src/test/java/use_case/FetchImage/FetchImageInteractorTest.java
@@ -3,8 +3,8 @@ package use_case.FetchImage;
 import data_access.SimpleImage;
 import entity.CommonImage;
 import data_access.InMemoryFetchImageDataAccessObject;
-import interface_adapter.image.FetchImagePresenter;
-import interface_adapter.image.FetchImageViewModel;
+import interface_adapter.fetch_image.FetchImagePresenter;
+import interface_adapter.fetch_image.FetchImageViewModel;
 import org.junit.jupiter.api.Test;
 import use_case.fetch_image.FetchImageInputData;
 import use_case.fetch_image.FetchImageInteractor;
